+++ src/analyze/src/components/analyze/AnalyzeSettings.vue
@@ -1,225 +1,138 @@
 <script>
-import { mapGetters, mapState } from 'vuex';
-import store from '@/store';
+import { mapState, mapGetters, mapActions } from 'vuex';
+import ConnectorLogo from '@/components/generic/ConnectorLogo';
+import ConnectorSettings from '@/components/pipelines/ConnectorSettings';
+
 
 export default {
+  name: 'AnalyzeSettings',
   data() {
     return {
-      connectionName: '',
-      connectionDatabase: '',
-      connectionSchema: '',
-      connectionDialect: '',
-      connectionHost: '',
-      connectionPort: '',
-      connectionUsername: '',
-      connectionPassword: '',
-      connectionSqlitePath: '',
+      connectionName: null,
     };
   },
-  beforeRouteEnter(to, from, next) {
-    store.dispatch('settings/getSettings')
-      .then(next)
-      .catch(() => {
-        next(from.path);
-      });
+  components: {
+    ConnectorLogo,
+    ConnectorSettings,
+  },
+  created() {
+    this.$store.dispatch('plugins/getAllPlugins');
+    this.$store.dispatch('plugins/getInstalledPlugins');
+  },
+  beforeDestroy() {
+    this.$store.dispatch('configuration/clearConnectionInFocusConfiguration');
   },
   computed: {
-    ...mapState('settings', [
-      'settings',
+    ...mapGetters('plugins', [
+      'getIsPluginInstalled',
+      'getIsInstallingPlugin',
     ]),
-    ...mapGetters('settings', [
-      'hasConnections',
-      'isConnectionDialectSqlite',
+    ...mapGetters('configuration', [
+      'getHasValidConfigSettings',
     ]),
+    ...mapState('configuration', [
+      'connectionInFocusConfiguration',
+    ]),
+    ...mapState('plugins', [
+      'plugins',
+      'installedPlugins',
+    ]),
+    connection() {
+      const targetConnection = this.installedPlugins.connections
+        ? this.installedPlugins.connections.find(item => item.name === this.connectionName)
+        : null;
+      return targetConnection || {};
+    },
+    configSettings() {
+      return this.connection.config
+        ? Object.assign(this.connection.config, this.connectionInFocusConfiguration)
+        : this.connectionInFocusConfiguration;
+    },
     isSaveable() {
-      // TODO proper validation
-      const dialectCondition = this.isConnectionDialectSqlite(this.connectionDialect)
-        ? this.connectionSqlitePath.length > 0
-        : true;
-      const val = dialectCondition &&
-        this.connectionName.length > 0 &&
-        this.connectionDatabase.length > 0 &&
-        this.connectionSchema.length > 0 &&
-        this.connectionDialect.length > 0 &&
-        this.connectionHost.length > 0 &&
-        this.connectionPort.length > 0 &&
-        this.connectionUsername.length > 0 &&
-        this.connectionPassword.length > 0;
-      return val;
+      const isInstalling = this.getIsInstallingPlugin('connections', this.connectionName);
+      const isInstalled = this.getIsPluginInstalled('connections', this.connectionName);
+      return !isInstalling && isInstalled;
     },
   },
   methods: {
-    saveConnection() {
-      this.$store.dispatch('settings/saveConnection', {
-        name: this.connectionName,
-        database: this.connectionDatabase,
-        schema: this.connectionSchema,
-        dialect: this.connectionDialect,
-        host: this.connectionHost,
-        port: this.connectionPort,
-        username: this.connectionUsername,
-        password: this.connectionPassword,
-        path: this.connectionSqlitePath,
+    ...mapActions('configuration', [
+      'getConnectionConfiguration',
+    ]),
+    configureConnection(connection) {
+      this.$store.dispatch('plugins/addPlugin', {
+        pluginType: 'connections',
+        name: connection,
+      }).then(() => {
+        this.connectionName = connection;
+        this.getConnectionConfiguration(connection);
       });
-      this.connectionName = '';
-      this.connectionDatabase = '';
-      this.connectionSchema = '';
-      this.connectionDialect = '';
-      this.connectionHost = '';
-      this.connectionPort = '';
-      this.connectionUsername = '';
-      this.connectionPassword = '';
-      this.connectionSqlitePath = '';
     },
-    deleteConnection(connection) {
-      this.$store.dispatch('settings/deleteConnection', connection);
+    saveConfig() {
+      this.$store.dispatch('configuration/savePluginConfiguration', {
+        type: 'connections',
+        name: this.connection.name,
+        config: this.configSettings.config,
+      });
     },
   },
 };
 </script>
 
 <template>
-  <section>
-    <div class="columns">
-      <div class="column">
-
-        <div class="content">
-          <h2>Connection Settings</h2>
-          <p v-if="!hasConnections">No Database Connections</p>
-          <div class="columns is-multiline is-mobile">
-            <div
-              class="column is-half"
-              v-for="connection in settings.connections"
-              :key="connection.host"
-            >
-              <div class="card">
-                <header class="card-header">
-                  <p class="card-header-title">{{connection.name}}</p>
-                </header>
-                <div class="card-content">
-                  <div class="content">
-                    <p>
-                      <strong>Dialect</strong>
-                      <span class="is-pulled-right">{{connection.dialect}}</span>
-                    </p>
-                    <div v-if="!isConnectionDialectSqlite(connection.dialect)">
-                      <p>
-                        <strong>Port</strong>
-                        <span class="is-pulled-right">{{connection.port}}</span>
-                      </p>
-                      <p>
-                        <strong>Username</strong>
-                        <span class="is-pulled-right">{{connection.username}}</span>
-                      </p>
-                      <p>
-                        <strong>Host</strong>
-                        <span
-                          class="ellipsis is-pulled-right"
-                          :title="connection.host"
-                        >{{connection.host}}</span>
-                      </p>
-                    </div>
-                    <div v-if="isConnectionDialectSqlite(connection.dialect)">
-                      <p>
-                        <strong>Path</strong>
-                        <span
-                          class="ellipsis is-pulled-right"
-                          :title="connection.path"
-                        >{{connection.path}}</span>
-                      </p>
-                    </div>
-                  </div>
+  <section class="columns">
+    <div class="column is-one-third">
+      <h2 class="title is-5">Available Connections</h2>
+      <div class="tile is-ancestor is-flex is-flex-column">
+        <div class="tile is-parent is-flex-no-grow"
+             v-for="(connection, index) in plugins.connections"
+             :key="`${connection}-${index}`">
+          <div class="tile level box">
+            <div class="level-left">
+              <div class="level-item is-flex-column has-text-left">
+                <ConnectorLogo class="connector-logo"
+                               :connector='connection'
+                               :is-grayscale='!getIsPluginInstalled("connections", connection)' />
+              </div>
+            </div>
+            <div class="level-right">
+              <div class="level-item content is-small is-flex-column">
+                <p class="is-uppercase has-text-weight-bold">{{ connection }}</p>
+                <div class="buttons are-small">
+                  <a class="button is-interactive-primary flex-grow-1"
+                     @click="configureConnection(connection)">Configure</a>
                 </div>
-                <footer class="card-footer">
-                  <a
-                    href="#"
-                    class="card-footer-item is-danger"
-                    @click.prevent="deleteConnection(connection)"
-                  >Delete Connection</a>
-                </footer>
               </div>
             </div>
           </div>
         </div>
+      </div>
+    </div>
 
-        <div class="content">
-          <h3 class="title">Add Analytics Connection</h3>
-          <div class="field is-grouped">
-            <div class="control is-expanded">
-              <input class="input" type="text" placeholder="Name" v-model="connectionName">
-            </div>
-            <div class="control">
-              <div class="select">
-                <select v-model="connectionDialect">
-                  <option value disabled selected>Dialect</option>
-                  <option value="postgresql">PostgreSQL</option>
-                  <option value="mysql">MySQL</option>
-                  <option value="sqlite">SQLite</option>
-                </select>
-              </div>
-            </div>
-          </div>
-
-          <div class="field" v-if="!isConnectionDialectSqlite(connectionDialect)">
-            <div class="field is-grouped">
-              <p class="control is-expanded">
-                <input class="input" type="text" placeholder="Host" v-model="connectionHost">
-              </p>
-              <p class="control">
-                <input class="input" type="text" placeholder="Port" v-model="connectionPort">
-              </p>
-            </div>
-
-            <div class="field is-grouped">
-              <p class="control is-expanded">
-                <input class="input" type="text" placeholder="Database" v-model="connectionDatabase">
-              </p>
-              <p class="control">
-                <input class="input" type="text" placeholder="Schema" v-model="connectionSchema">
-              </p>
-            </div>
-
-            <div class="field is-grouped">
-              <p class="control is-expanded">
-                <input class="input" type="text" placeholder="Username" v-model="connectionUsername">
-              </p>
-              <p class="control is-expanded">
-                <input
-                  class="input"
-                  type="password"
-                  placeholder="Password"
-                  v-model="connectionPassword"
-                >
-              </p>
-            </div>
-          </div>
-
-          <div class="field" v-if="isConnectionDialectSqlite(connectionDialect)">
-            <div class="field is-grouped">
-              <p class="control is-expanded">
-                <input
-                  class="input"
-                  type="text"
-                  placeholder="Path to SQLite file"
-                  v-model="connectionSqlitePath"
-                >
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div class="buttons is-right">
-          <button
-            class="button is-interactive-primary"
-            :disabled="!isSaveable"
-            @click.prevent="saveConnection"
-          >Save</button>
-        </div>
-
+    <div class="column" rel="container" v-if="configSettings">
+      <h2 class="title is-5">Configuration</h2>
+      <ConnectorSettings v-if='connectionName'
+                         class="box"
+                         :config-settings='configSettings'>
+        <section class="field buttons is-right"
+                 slot="bottom">
+          <button class='button is-interactive-primary'
+                  :disabled='!isSaveable'
+                  @click.prevent="saveConfig">Save</button>
+        </section>
+      </ConnectorSettings>
+      <div v-else
+           class="box">
+        <p>
+          Please select a connector to configure.
+        </p>
       </div>
     </div>
   </section>
 </template>
 
 <style lang="scss">
+.connector-logo {
+  max-height: 48px;
+  object-fit: scale-down;
+}
 </style>