+++ src/analyze/src/store/modules/designs.js
@@ -108,6 +108,7 @@ const helpers = {
       order,
       limit: state.limit,
       filters: state.filters,
+      dialect: state.dialect,
     };
   },
 };
@@ -231,14 +232,6 @@ const actions = {
         commit('setDesign', response.data);
       });
 
-    sqlApi.getDialect(model)
-      .then((response) => {
-        commit('setConnectionDialect', response.data);
-      })
-      .catch((e) => {
-        commit('setSqlErrorMessage', e);
-      });
-
     reportsApi.loadReports()
       .then((response) => {
         state.reports = response.data;
@@ -325,6 +318,7 @@ const actions = {
     commit('setLimit', limit);
   },
 
+  // TODO: remove and use `mapMutations`
   setChartType({ commit }, chartType) {
     commit('setChartType', chartType);
   },
@@ -333,7 +327,7 @@ const actions = {
     this.dispatch('designs/resetErrorMessage');
     state.loadingQuery = !!run;
 
-    const queryPayload = load || helpers.getQueryPayloadFromDesign();
+    const queryPayload = Object.assign({}, helpers.getQueryPayloadFromDesign(), load);
     const postData = Object.assign({ run }, queryPayload);
     sqlApi
       .getSql(state.currentModel, state.currentDesign, postData)
@@ -463,6 +457,7 @@ const mutations = {
     state.currentDesign = report.design;
     state.chartType = report.chartType;
     state.limit = report.queryPayload.limit;
+    state.dialect = report.queryPayload.dialect;
 
     // UI selected state adornment helpers for columns, aggregates, filters, joins, & timeframes
     const baseTable = state.design.related_table;
@@ -560,10 +555,6 @@ const mutations = {
     state.currentSQL = results.sql;
   },
 
-  setConnectionDialect(_, results) {
-    state.dialect = results.connection_dialect;
-  },
-
   setQueryResults(_, results) {
     state.results = results.results;
     state.keys = results.keys;
@@ -604,6 +595,10 @@ const mutations = {
 
   setLimit(_, limit) {
     state.limit = limit;
+  },
+
+  setDialect(_, dialect) {
+    state.dialect = dialect;
   },
 };