+++ src/meltano/cli/__init__.py
@@ -12,6 +12,7 @@ from . import (
     schedule,
     select,
     repl,
+    config,
 )